requirements.txt
Django==5.1.7
asgiref==3.8.1
django-ckeditor==6.7.2
django-js-asset==3.1.2
pymongo==4.11.3
bleach==6.2.0
requests==2.32.3
sqlparse==0.5.3